+++ src/components/Admin/Content/ManageUser.js
@@ -1,9 +1,24 @@
 import ModalCreateUser from './ModalCreateUser';
 import './ManageUser.scss';
-import { useState } from 'react';
 import TableUser from './TableUser';
+import { useEffect, useState } from 'react';
+import { getAllUsers } from '../../../services/apiService';
 const ManageUser = (props) => {
   const [showModalCreateUser, setSowModalCreateUser] = useState(false);
+
+  const [listUsers, setListUsers] = useState([]);
+  //ComponentDidMount
+  useEffect(() => {
+    fetchListUsers();
+  }, []);
+
+  const fetchListUsers = async () => {
+    let res = await getAllUsers();
+    if (res.EC === 0) {
+      setListUsers(res.DT);
+    }
+  };
+
   return (
     <div className="manage-user-container">
       <div className="title">Manage User</div>
@@ -14,9 +29,9 @@ const ManageUser = (props) => {
           </button>
         </div>
         <div className="table-user-container">
-          <TableUser />
+          <TableUser listUsers={listUsers} />
         </div>
-        <ModalCreateUser show={showModalCreateUser} setShow={setSowModalCreateUser} />
+        <ModalCreateUser show={showModalCreateUser} setShow={setSowModalCreateUser} fetchListUsers={fetchListUsers} />
       </div>
     </div>
   );
