+++ src/components/Admin/Content/ModalCreateUser.js
@@ -44,11 +44,11 @@ const ModalCreateUser = (props) => {
 
   const handleSubmitCreateUser = async () => {
     // validate
-    // const isValidEmail = validateEmail(email);
-    // if (!isValidEmail) {
-    //   toast.error('Invalid email');
-    //   return;
-    // }
+    const isValidEmail = validateEmail(email);
+    if (!isValidEmail) {
+      toast.error('Invalid email');
+      return;
+    }
     if (!password) {
       toast.error('Invalid password');
       return;
@@ -58,6 +58,7 @@ const ModalCreateUser = (props) => {
     if (data && data.EC === 0) {
       toast.success(data.EM);
       handleClose();
+      await props.fetchListUsers();
     }
 
     if (data && data.EC !== 0) {
